--- Domain/src/main/java/com/dingdong/domain/domains/idcard/domain/UserInfo.java
@@ -1,19 +1,0 @@
-package com.dingdong.domain.domains.idcard.domain;
-
-
-import javax.persistence.Embeddable;
-import javax.persistence.Embedded;
-import javax.validation.constraints.NotNull;
-import lombok.Getter;
-
-@Embeddable
-@Getter
-public class UserInfo {
-    @NotNull private Long userId;
-
-    @NotNull private String nickname;
-
-    @NotNull private String aboutMe;
-
-    @Embedded @NotNull private Character character;
-}
